+++ tests/Console/ModelsCommand/migrations/____custom_casts_table.php
@@ -12,6 +12,7 @@ class CustomCastsTable extends Migration
     {
         Schema::create('custom_casts', static function (Blueprint $table) {
             $table->string('casted_property_with_return_type');
+            $table->string('casted_property_with_static_return_type');
             $table->string('casted_property_with_return_docblock');
             $table->string('casted_property_with_return_docblock_fqn');
             $table->string('casted_property_with_return_primitive');
@@ -22,9 +23,14 @@ class CustomCastsTable extends Migration
             $table->string('casted_property_with_param');
             $table->string('casted_property_with_static_return_docblock');
             $table->string('casted_property_with_this_return_docblock');
+            $table->string('casted_property_with_castable');
+            $table->string('casted_property_with_anonymous_cast');
+            $table->string('casted_property_without_return_type');
             $table->string('extended_casted_property_with_static_return_docblock');
             $table->string('extended_casted_property_with_this_return_docblock');
             $table->string('casted_property_with_static_return_docblock_and_param');
+            $table->string('casted_property_with_collection_generics');
+            $table->string('casted_property_with_nullable_collection_generics');
         });
     }
 }